moving api functions to correct location
