--- clients/linux-app/Auth/PassFunctions.py
@@ -1,13 +1,0 @@
-import bcrypt
-import mysql.connector
-
-
-def hash_password(password: str):
-    # Generate a random salt
-    salt = bcrypt.gensalt()
-
-    # Hash the password with the salt
-    hashed_password = bcrypt.hashpw(password.encode('utf-8'), salt)
-
-    # Return the salt and the hashed password
-    return salt, hashed_password
